Make app folder standalone
