+++ core/apps/ame/src/app/components/loading/loading.component.ts
@@ -18,8 +18,10 @@ import {forkJoin, Observable, of, Subscription, switchMap, take} from 'rxjs';
 import {NgOptimizedImage} from '@angular/common';
 
 @Component({
+  standalone: true,
   templateUrl: 'loading.component.html',
   styleUrls: ['loading.component.scss'],
+  imports: [NgOptimizedImage],
 })
 export class LoadingComponent implements AfterViewInit, OnDestroy {
   private subscription = new Subscription();
